backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
python-dotenv
requests
gspread>=6.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
